fix: removed circular dependencies
